move find_ancestors_by_hierarchial_name() to its natural home
